Split up WebhooksController into Webhooks:: namespace
